pyproject.toml
[project]
name = "salariez"
version = "0.1.0"
description = "Salaries Management Web Application"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-extensions>=3.2.3",
    "django==5.0.1",
    "django-filter>=24.3",
    "django-cotton>=1.3.0",
    "django-htmx>=1.21.0",
    "django-widget-tweaks>=1.5.0",
    "django-import-export[xlsx]>=4.3.1",
    "django-mptt>=0.16.0",
    "django-braces>=1.16.0",
    "djangoql>=0.18.1",
    "djangorestframework>=3.15.2",
    "djangorestframework-simplejwt>=5.3.1",
    "mysqlclient>=2.2.6",
    "python-dotenv>=1.0.1",
    "pillow>=11.0.0",
    "whitenoise>=6.8.2",
    "blacknoise>=1.1.0",
]

[dependency-groups]
dev = [
    "django-silk>=5.3.1",
    "pytest>=8.3.3",
    "selectolax>=0.3.26",
]
